refactoring of app.js, main navbar components
