+++ react-app/src/components/MainNavBar/index.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+import NavBar from './NavBar';
+
+const MainNavBar = () => <NavBar />;
+
+export default MainNavBar;
